feat: render working set entry actions based on entry state
