+++ src/vs/workbench/contrib/chat/browser/chatEditingActions.ts
@@ -1,0 +1,192 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Codicon } from '../../../../base/common/codicons.js';
+import { URI } from '../../../../base/common/uri.js';
+import { ServicesAccessor } from '../../../../editor/browser/editorExtensions.js';
+import { localize, localize2 } from '../../../../nls.js';
+import { Action2, MenuId, registerAction2 } from '../../../../platform/actions/common/actions.js';
+import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextkey.js';
+import { EditorActivation } from '../../../../platform/editor/common/editor.js';
+import { IEditorService } from '../../../services/editor/common/editorService.js';
+import { CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME, IChatEditingService, IChatEditingSession, WorkingSetEntryState, chatEditingResourceContextKey, chatEditingWidgetFileStateContextKey, decidedChatEditingResourceContextKey } from '../common/chatEditingService.js';
+import { IChatWidget, IChatWidgetService } from './chat.js';
+
+abstract class WorkingSetAction extends Action2 {
+	run(accessor: ServicesAccessor, ...args: any[]) {
+		const chatEditingService = accessor.get(IChatEditingService);
+		const currentEditingSession = chatEditingService.currentEditingSession;
+		if (!currentEditingSession) {
+			return;
+		}
+
+		const chatWidget = accessor.get(IChatWidgetService).lastFocusedWidget;
+		const uris: URI[] = [];
+		if (URI.isUri(args[0])) {
+			uris.push(args[0]);
+		} else if (chatWidget) {
+			uris.push(...chatWidget.input.selectedElements);
+		}
+		if (!uris.length) {
+			return;
+		}
+
+		return this.runWorkingSetAction(accessor, currentEditingSession, chatWidget, ...uris);
+	}
+
+	abstract runWorkingSetAction(accessor: ServicesAccessor, currentEditingSession: IChatEditingSession, chatWidget: IChatWidget | undefined, ...uris: URI[]): any;
+}
+
+registerAction2(class OpenFileAction extends WorkingSetAction {
+	constructor() {
+		super({
+			id: 'chatEditing.openFile',
+			title: localize2('open.file', 'Open File'),
+			icon: Codicon.goToFile,
+			menu: [{
+				id: MenuId.ChatEditingSessionWidgetToolbar,
+				when: ContextKeyExpr.equals(chatEditingWidgetFileStateContextKey.key, WorkingSetEntryState.Edited),
+				order: 0,
+				group: 'navigation'
+			}],
+		});
+	}
+
+	async runWorkingSetAction(accessor: ServicesAccessor, currentEditingSession: IChatEditingSession, chatWidget: IChatWidget, ...uris: URI[]): Promise<void> {
+		const editorService = accessor.get(IEditorService);
+		await Promise.all(uris.map((uri) => editorService.openEditor({ resource: uri, options: { pinned: true, activation: EditorActivation.ACTIVATE } })));
+	}
+});
+
+registerAction2(class AcceptAction extends WorkingSetAction {
+	constructor() {
+		super({
+			id: 'chatEditing.acceptFile',
+			title: localize2('accept.file', 'Accept'),
+			icon: Codicon.check,
+			menu: [{
+				when: ContextKeyExpr.and(ContextKeyExpr.equals('resourceScheme', CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME), ContextKeyExpr.notIn(chatEditingResourceContextKey.key, decidedChatEditingResourceContextKey.key)),
+				id: MenuId.MultiDiffEditorFileToolbar,
+				order: 0,
+				group: 'navigation',
+			}, {
+				id: MenuId.ChatEditingSessionWidgetToolbar,
+				when: ContextKeyExpr.equals(chatEditingWidgetFileStateContextKey.key, WorkingSetEntryState.Edited),
+				order: 2,
+				group: 'navigation'
+			}],
+		});
+	}
+
+	async runWorkingSetAction(accessor: ServicesAccessor, currentEditingSession: IChatEditingSession, chatWidget: IChatWidget, ...uris: URI[]): Promise<void> {
+		await currentEditingSession.accept(...uris);
+	}
+});
+
+registerAction2(class DiscardAction extends WorkingSetAction {
+	constructor() {
+		super({
+			id: 'chatEditing.discardFile',
+			title: localize2('discard.file', 'Discard'),
+			icon: Codicon.discard,
+			menu: [{
+				when: ContextKeyExpr.and(ContextKeyExpr.equals('resourceScheme', CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME), ContextKeyExpr.notIn(chatEditingResourceContextKey.key, decidedChatEditingResourceContextKey.key)),
+				id: MenuId.MultiDiffEditorFileToolbar,
+				order: 0,
+				group: 'navigation',
+			}, {
+				id: MenuId.ChatEditingSessionWidgetToolbar,
+				when: ContextKeyExpr.equals(chatEditingWidgetFileStateContextKey.key, WorkingSetEntryState.Edited),
+				order: 1,
+				group: 'navigation'
+			}],
+		});
+	}
+
+	async runWorkingSetAction(accessor: ServicesAccessor, currentEditingSession: IChatEditingSession, chatWidget: IChatWidget, ...uris: URI[]): Promise<void> {
+		await currentEditingSession.reject(...uris);
+	}
+});
+
+export class ChatEditingAcceptAllAction extends Action2 {
+	static readonly ID = 'chatEditing.acceptAllFiles';
+	static readonly LABEL = localize('accept.allFiles', 'Accept All');
+
+	constructor() {
+		super({
+			id: ChatEditingAcceptAllAction.ID,
+			title: ChatEditingAcceptAllAction.LABEL,
+			// icon: Codicon.goToFile,
+			menu: {
+				when: ContextKeyExpr.equals('resourceScheme', CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME),
+				id: MenuId.EditorTitle,
+				order: 0,
+				group: 'navigation',
+			},
+		});
+	}
+
+	async run(accessor: ServicesAccessor, ...args: any[]): Promise<void> {
+		const chatEditingService = accessor.get(IChatEditingService);
+		const currentEditingSession = chatEditingService.currentEditingSession;
+		if (!currentEditingSession) {
+			return;
+		}
+		await currentEditingSession.accept();
+	}
+}
+registerAction2(ChatEditingAcceptAllAction);
+
+export class ChatEditingDiscardAllAction extends Action2 {
+	static readonly ID = 'chatEditing.discardAllFiles';
+	static readonly LABEL = localize('discard.allFiles', 'Discard All');
+
+	constructor() {
+		super({
+			id: ChatEditingDiscardAllAction.ID,
+			title: ChatEditingDiscardAllAction.LABEL,
+			// icon: Codicon.goToFile,
+			menu: {
+				when: ContextKeyExpr.equals('resourceScheme', CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME),
+				id: MenuId.EditorTitle,
+				order: 0,
+				group: 'navigation',
+			},
+		});
+	}
+
+	async run(accessor: ServicesAccessor, ...args: any[]): Promise<void> {
+		const chatEditingService = accessor.get(IChatEditingService);
+		const currentEditingSession = chatEditingService.currentEditingSession;
+		if (!currentEditingSession) {
+			return;
+		}
+		await currentEditingSession.reject();
+	}
+}
+registerAction2(ChatEditingDiscardAllAction);
+
+export class ChatEditingShowChangesAction extends Action2 {
+	static readonly ID = 'chatEditing.openDiffs';
+	static readonly LABEL = localize('chatEditing.openDiffs', 'Open Diffs');
+
+	constructor() {
+		super({
+			id: ChatEditingShowChangesAction.ID,
+			title: ChatEditingShowChangesAction.LABEL,
+			f1: false
+		});
+	}
+
+	async run(accessor: ServicesAccessor, ...args: any[]): Promise<void> {
+		const chatEditingService = accessor.get(IChatEditingService);
+		const currentEditingSession = chatEditingService.currentEditingSession;
+		if (!currentEditingSession) {
+			return;
+		}
+		await currentEditingSession.show();
+	}
+}
+registerAction2(ChatEditingShowChangesAction);
