+++ src/vs/workbench/contrib/chat/browser/chatInputPart.ts
@@ -64,7 +64,7 @@ import { AccessibilityCommandId } from '../../accessibility/common/accessibility
 import { getSimpleCodeEditorWidgetOptions, getSimpleEditorOptions, setupSimpleEditorSelectionStyling } from '../../codeEditor/browser/simpleEditorOptions.js';
 import { ChatAgentLocation, IChatAgentService } from '../common/chatAgents.js';
 import { CONTEXT_CHAT_INPUT_CURSOR_AT_TOP, CONTEXT_CHAT_INPUT_HAS_FOCUS, CONTEXT_CHAT_INPUT_HAS_TEXT, CONTEXT_IN_CHAT_INPUT } from '../common/chatContextKeys.js';
-import { ChatEditingSessionState, IChatEditingSession, ModifiedFileEntryState } from '../common/chatEditingService.js';
+import { ChatEditingSessionState, IChatEditingSession, WorkingSetEntryState } from '../common/chatEditingService.js';
 import { IChatRequestVariableEntry } from '../common/chatModel.js';
 import { IChatFollowup } from '../common/chatService.js';
 import { IChatResponseViewModel } from '../common/chatViewModel.js';
@@ -879,6 +879,7 @@ export class ChatInputPart extends Disposable implements IHistoryNavigationWidge
 			if (attachment.isFile && URI.isUri(attachment.value) && !modifiedFiles.has(attachment.value)) {
 				entries.unshift({
 					reference: attachment.value,
+					state: WorkingSetEntryState.Attached,
 					kind: 'reference',
 				});
 				modifiedFiles.add(attachment.value);
@@ -888,6 +889,7 @@ export class ChatInputPart extends Disposable implements IHistoryNavigationWidge
 			if (!modifiedFiles.has(file)) {
 				entries.unshift({
 					reference: file,
+					state: WorkingSetEntryState.Attached,
 					kind: 'reference',
 				});
 			}
@@ -940,16 +942,19 @@ export class ChatInputPart extends Disposable implements IHistoryNavigationWidge
 			this._chatEditList = this._chatEditsListPool.get();
 			const list = this._chatEditList.object;
 			this._chatEditsDisposables.add(this._chatEditList);
+			this._chatEditsActionsDisposables.add(list.onDidFocus(() => {
+				this._onDidFocus.fire();
+			}));
 			this._chatEditsDisposables.add(list.onDidOpen((e) => {
 				if (e.element?.kind === 'reference' && URI.isUri(e.element.reference)) {
 					const modifiedFileUri = e.element.reference;
 					const editedFile = chatEditingSession.entries.get().find((e) => e.modifiedURI.toString() === modifiedFileUri.toString());
-					if (editedFile?.state.get() === ModifiedFileEntryState.Undecided) {
+					if (editedFile?.state.get() === WorkingSetEntryState.Edited) {
 						void this.editorService.openEditor({
 							original: { resource: URI.from(editedFile.originalURI, true) },
 							modified: { resource: URI.from(editedFile.modifiedURI, true) },
 						});
-					} else if (editedFile) {
+					} else {
 						void this.editorService.openEditor({ resource: modifiedFileUri });
 					}
 				}
@@ -971,7 +976,7 @@ export class ChatInputPart extends Disposable implements IHistoryNavigationWidge
 			dom.clearNode(actionsContainer);
 			const actionsContainerRight = actionsContainer.querySelector('.chat-editing-session-actions-group') as HTMLElement ?? $('.chat-editing-session-actions-group');
 
-			if (chatEditingSession.entries.get().find((e) => e.state.get() === ModifiedFileEntryState.Undecided)) {
+			if (chatEditingSession.entries.get().find((e) => e.state.get() === WorkingSetEntryState.Edited)) {
 				// Don't show Accept All / Discard All actions if user already selected Accept All / Discard All
 				const actions = [];
 				actions.push(
