+++ src/vs/workbench/contrib/chat/common/chatEditingService.ts
@@ -7,6 +7,8 @@ import { Event } from '../../../../base/common/event.js';
 import { IObservable, ITransaction } from '../../../../base/common/observable.js';
 import { URI } from '../../../../base/common/uri.js';
 import { TextEdit } from '../../../../editor/common/languages.js';
+import { localize } from '../../../../nls.js';
+import { RawContextKey } from '../../../../platform/contextkey/common/contextkey.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
 import { IChatResponseModel } from './chatModel.js';
 
@@ -41,8 +43,9 @@ export interface IChatEditingSession {
 	stop(): Promise<void>;
 }
 
-export const enum ModifiedFileEntryState {
-	Undecided,
+export const enum WorkingSetEntryState {
+	Attached,
+	Edited,
 	Accepted,
 	Rejected
 }
@@ -50,7 +53,7 @@ export const enum ModifiedFileEntryState {
 export interface IModifiedFileEntry {
 	readonly originalURI: URI;
 	readonly modifiedURI: URI;
-	readonly state: IObservable<ModifiedFileEntryState>;
+	readonly state: IObservable<WorkingSetEntryState>;
 	accept(transaction: ITransaction | undefined): Promise<void>;
 	reject(transaction: ITransaction | undefined): Promise<void>;
 }
@@ -65,3 +68,10 @@ export const enum ChatEditingSessionState {
 	Idle = 2,
 	Disposed = 3
 }
+
+export const CHAT_EDITING_MULTI_DIFF_SOURCE_RESOLVER_SCHEME = 'chat-editing-multi-diff-source';
+
+export const chatEditingWidgetFileStateContextKey = new RawContextKey<WorkingSetEntryState>('chatEditingWidgetFileState', undefined, localize('chatEditingWidgetFileState', "The current state of the file in the chat editing widget"));
+export const decidedChatEditingResourceContextKey = new RawContextKey<string[]>('decidedChatEditingResource', []);
+export const chatEditingResourceContextKey = new RawContextKey<string | undefined>('chatEditingResource', undefined);
+export const inChatEditingSessionContextKey = new RawContextKey<boolean | undefined>('inChatEditingSession', undefined);
